backend/requirements.txt
# === Web Framework ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# === Database ===
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
geoalchemy2==0.15.2

# === Geospatial ===
shapely==2.0.6
geojson-pydantic==1.1.2
geopandas==1.0.1
pyproj==3.7.0

# === Data Processing ===
dask[distributed]==2024.12.1
pandas==2.2.3

# === S3 / MinIO ===
boto3==1.35.81

# === Migrations ===
alembic==1.14.0

# === Config ===
pydantic-settings==2.7.0

# === ML Inference ===
onnxruntime==1.19.2
scikit-learn==1.4.2
numpy==1.26.4

# === Utilities ===
orjson==3.10.12
httpx==0.28.1
